[ADD] Suggestion for future
What to do?
Reshape the GA: instead of sending a list of GameObjects, send only a list of the bird brains.

Why?
Due to an unknown issue, after a restart of the game birds seem to be inactive. So the idea is to recreate them by default, as to not reset them by hand.